+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 When stuff changes, it's described here.
 
+## 2015-01-08
+ * Moved shipping provider to config so it can be easily changed
+
 ## 2015-01-07
  * Refactored logic into separate prototype functions to make it easier to extend
  * Added php test (e.g. php test/test.php) to show how to communicate with the service via php